remove intermediary type and helper
