+++ src/slices/components/marketFilter/index.ts
@@ -1,3 +1,4 @@
 export { default } from "./marketFilter";
 export * from "./marketFilter";
 export { SDG_GROUPS } from "./constants";
+export * from "./types";
